shifted all function of writeFile to utils
