+++ src/fileSystem.js
@@ -11,63 +11,7 @@ import {
 } from 'fs';
 import isValid from 'is-valid-path';
 
-/**
- * fixFileName()
- * @param {string} fileName
- */
-const fixFileName = (fileName) => {
-  let correctedFileName;
-
-  correctedFileName = fileName.replace(/ /g, '_'); // Replace space with underscore
-  correctedFileName = fileName.toLowerCase(); // Maintain Uniformity
-
-  return correctedFileName;
-};
-
-/**
- * for stripSymbols()
- * @param {Object} data a json object
- *
- */
-const stripSymbols = (data) => {
-  let dataStr = JSON.stringify(data);
-
-  const replaceList = [
-    ['/{"/g', '{ "'],
-    ['/{"/g', '{ " '],
-    ['/},{/g', ' },\n{'],
-    ['/":/g', '": '],
-    ['/,"/g', ',\n "'],
-  ];
-
-  replaceList.forEach((replacer) => {
-    dataStr = dataStr.replace(replacer[0], replacer[1]);
-  });
-
-  return dataStr;
-};
-
-/**
- * getFileName()
- * @param {string} file
- * @param {Object} fileData
- * @param {var} flag
- * @param {var} index
- */
-// @TODO if we use fileData.name - why we didn't just pass it here?
-const getFileName = (file, fileData, flag, index) => {
-  let fileName;
-  if (flag === 1) {
-    // for example: 23-someJsonFile.json
-    fileName = `${index}-${file}`;
-  } else {
-    // for example: someValueOfName.json
-    fileName = `${fileData.name}.json`;
-  }
-
-  fileName = fixFileName(fileName);
-  return fileName;
-};
+const utils = require('./utils');
 
 /**
  * Write in file
@@ -87,7 +31,7 @@ const write = (path, data) => new Promise((resolve) => {
   if (typeof data === 'string') {
     dataStr = data;
   } else {
-    dataStr = stripSymbols(data);
+    dataStr = utils.stripSymbols(data);
   }
 
   writeFile(path, dataStr, (err) => {
@@ -148,7 +92,7 @@ const save = (folderNamePath, file, fileData, flag) => {
     // @TODO long line, I have feeling that it can be improved
     // - we just need to find a better way to
     // rewrite a getFileName method
-    const fileName = getFileName(file, fileData[i], flag, i);
+    const fileName = utils.getFileName(file, fileData[i], flag, i);
 
     const elementPath = `${folderNamePath}/${fileName}`;
     const result = write(elementPath, fileData[i]);
@@ -206,14 +150,5 @@ const makeFolder = (path, file) => {
 };
 
 export {
-  write,
-  read,
-  save,
-  makeFolder,
-  isFolderExists,
-  dirSync,
-  syncStats,
-  stripSymbols,
-  getFileName,
-  fixFileName,
+  write, read, save, makeFolder, isFolderExists, dirSync, syncStats,
 };
